+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v2.0
 - Erich Matos Viegas: fixed chat messages
 - Erich Matos Viegas: you can custom everything
+- Settings Refactor 
 
 # v1.9
 - v13